requirements.txt
streamlit
pymysql
tensorflow
requests
pandas
scikit-learn
konlpy
python-dotenv
bcrypt
